refactor: Refatorado classe do SiTef para não retornar métodos privados
