+++ src/utils/createPromise.ts
@@ -1,0 +1,10 @@
+export default <T = void>(handler: Function) => {
+  return new Promise<T>((resolve, reject) => {
+    try {
+      const value = handler();
+      resolve(value);
+    } catch (error) {
+      reject(error);
+    }
+  });
+};
